requirements.txt
ago==0.0.95
asgiref==3.6.0
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
boto3==1.26.123
botocore==1.29.123
bs4==0.0.1
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
coverage==7.2.3
cryptography==40.0.2
cssselect==1.2.0
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
docopt==0.6.2
dotmap==1.3.30
elastic-transport==8.4.0
elasticsearch==8.7.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
fingerprint==0.1.6
hjson==3.1.0
hurry.filesize==0.9
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.2.0
langdetect==1.0.9
lxml==4.9.2
news-please==1.5.33
newspaper3k==0.2.8
nltk==3.8.1
packaging==23.1
parsel==1.8.1
Pillow==9.5.0
pipreqs==0.4.13
plac==1.3.5
Protego==0.2.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyMySQL==1.0.3
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
queuelib==1.6.2
readability-lxml==0.8.1
regex==2023.3.23
requests==2.29.0
requests-file==1.5.1
s3transfer==0.6.0
Scrapy==2.8.0
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tinysegmenter==0.3
tldextract==3.4.1
tqdm==4.65.0
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.15
w3lib==2.1.1
warcio==1.7.4
yarg==0.1.9
zope.interface==6.0
